requirements.txt
pyqt6
psutil
pyserial
pyqtgraph